% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_catalog.R
\name{init_catalog}
\alias{init_catalog}
\title{Initialize a catalog from one or more sources}
\usage{
init_catalog(srcs)
}
\arguments{
\item{srcs}{a list of sources}
}
\description{
Initialize a catalog from one or more sources
}
