% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcs.R
\name{local_file_src}
\alias{local_file_src}
\title{Use a local file src}
\usage{
local_file_src(paths, extension, recurse = FALSE, read_fun = read.csv, ...)
}
\arguments{
\item{paths}{vector of paths to files and/or directories containing files}

\item{extension}{which file path extensions to look for}

\item{recurse}{whether to search recursively in directories}

\item{read_fun}{callable function capable of reading files}

\item{...}{arguments passed to \code{read_fun}}
}
\description{
Use a local file src
}
